Prune Objects --keep-threads
This adds an option to the prune_objects mix task.
The original way deleted all non-local public posts older than a certain time frame.
Here we add a different query which you can call using the option --keep-threads.

We query from the activities table all context id's where
    1. the newest activity with this context is still old
    2. none of the activities with this context is is local
    3. none of the activities with this context is bookmarked
and delete all objects with these contexts.

The idea is that posts with local activities (posts, replies, likes, repeats...) may be intersesting to keep.
Besides that, a post lives in a certain context (the thread), so we keep the whole thread as well.

Caveats:
* Quotes have a different context. Therefore, when someone quotes a post, it's possible the quoted post will still be deleted.
* Although undocumented (in docs/docs/administration/CLI_tasks/database.md/#prune-old-remote-posts-from-the-database), the 'normal' delete action still keeps old remote non-public posts. With this option we don't care about scope.
* I ran this on my instance, but directly on the DB. I still need to test to be sure that we don't get a time-out error or something.

Some statistics from explain analyse:
(cost=1402845.92..1933782.00 rows=3810907 width=62) (actual time=2562455.486..2562455.495 rows=0 loops=1)
 Planning Time: 505.327 ms
 Trigger for constraint chat_message_references_object_id_fkey: time=651939.797 calls=921740
 Trigger for constraint deliveries_object_id_fkey: time=52036.009 calls=921740
 Trigger for constraint hashtags_objects_object_id_fkey: time=20665.778 calls=921740
 Execution Time: 3287933.902 ms

+++ lib/mix/tasks/pleroma/database.ex
@@ -67,7 +67,8 @@ defmodule Mix.Tasks.Pleroma.Database do
       OptionParser.parse(
         args,
         strict: [
-          vacuum: :boolean
+          vacuum: :boolean,
+          keep_threads: :boolean
         ]
       )
 
@@ -77,24 +78,47 @@ defmodule Mix.Tasks.Pleroma.Database do
 
     Logger.info("Pruning objects older than #{deadline} days")
 
-    time_deadline =
-      NaiveDateTime.utc_now()
-      |> NaiveDateTime.add(-(deadline * 86_400))
+    if Keyword.get(options, :keep_threads) do
+      # We delete objects from threads where
+      # 1. the newest post is still old
+      # 2. none of the activities is local
+      # 3. none of the activities is bookmarked
+      delete_keep_threads_statement = """
+      delete
+      from public.objects o
+      where o.data ->> 'context' in (
+      select
+      a.data ->> 'context'
+      from public.activities a
+      left join public.bookmarks b on a.id = b.activity_id
+      group by (a.data ->> 'context'::text)
+      having max(a.updated_at) < now() - interval '#{deadline} day'
+      and not bool_or(a.local)
+      and max(b.id) is null
+      );
+      """
 
-    from(o in Object,
-      where:
-        fragment(
-          "?->'to' \\? ? OR ?->'cc' \\? ?",
-          o.data,
-          ^Pleroma.Constants.as_public(),
-          o.data,
-          ^Pleroma.Constants.as_public()
-        ),
-      where: o.inserted_at < ^time_deadline,
-      where:
-        fragment("split_part(?->>'actor', '/', 3) != ?", o.data, ^Pleroma.Web.Endpoint.host())
-    )
-    |> Repo.delete_all(timeout: :infinity)
+      Repo.query(delete_keep_threads_statement)
+    else
+      time_deadline =
+        NaiveDateTime.utc_now()
+        |> NaiveDateTime.add(-(deadline * 86_400))
+
+      from(o in Object,
+        where:
+          fragment(
+            "?->'to' \\? ? OR ?->'cc' \\? ?",
+            o.data,
+            ^Pleroma.Constants.as_public(),
+            o.data,
+            ^Pleroma.Constants.as_public()
+          ),
+        where: o.inserted_at < ^time_deadline,
+        where:
+          fragment("split_part(?->>'actor', '/', 3) != ?", o.data, ^Pleroma.Web.Endpoint.host())
+      )
+      |> Repo.delete_all(timeout: :infinity)
+    end
 
     prune_hashtags_query = """
     DELETE FROM hashtags AS ht
